Split geometry
